Cleaned up header work; Added resume for test;
