+++ src/interfaces/header.ts
@@ -1,6 +1,7 @@
 export interface MenuItem {
   title: string
   id?: string
+  url?: string
 }
 
 export interface SocialMediaLink {
